// lib/models/parkingEstablishment.ts
export interface ParkingEstablishment {
    access_info: string;
    accessibility: string;
    created_at: string;
    custom_access: string;
    custom_layout: string;
    dimensions: string;
    establishment_id: number;
    facilities: string;
    is24_7: boolean;
    latitude: number;
    lighting: string;
    longitude: number;
    name: string;
    nearby_landmarks: string;
    profile_id: number;
    space_layout: string;
    space_type: string;
    updated_at: string;
    uuid: string;
    verified: boolean;
}